requirements.txt
astroid==2.0.4
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pylint==2.1.1
python-ldap==3.1.0
six==1.11.0
wrapt==1.10.11
